+++ config_env.py
@@ -10,7 +10,7 @@ Security Features:
 - Configurable security settings
 - Development vs Production modes
 
-Author: Ahsan Iqbal
+Author: Just Ahsan
 Course: COM7033 - Secure Programming
 Institution: Leeds Trinity University
 """