+++ database/__init__.py
@@ -3,7 +3,7 @@ Database Package
 
 Contains database management modules for the Stroke Intelligence System.
 
-Author: Ahsan Iqbal
+Author: Just Ahsan
 Course: COM7033 - Secure Programming
 Institution: Leeds Trinity University